_Legacy_Archive_2026_01_30/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
watchdog==4.0.0
Pillow==10.2.0
numpy==1.26.4
trimesh==4.1.4
jinja2==3.1.3
python-multipart==0.0.9
requests==2.31.0
networkx==3.2.1
scipy==1.12.0
lxml==5.1.0
chardet==5.2.0
patool
